fix: compiler warnings and use generate cmake file (#1)

+++ src/include/nimble-steps-serialize/out_serialize.h
@@ -13,11 +13,11 @@ struct FldOutStream;
 int nbsStepsOutSerialize(struct FldOutStream* stream, const NbsSteps* steps);
 int nbsStepsOutSerializeFixedCountNoHeader(struct FldOutStream* stream, StepId startStepId, size_t redundancyCount,
                                            const NbsSteps* steps);
-int nbsStepsOutSerializeStep(const NimbleStepsOutSerializeLocalParticipants* participants, uint8_t* buf,
+ssize_t nbsStepsOutSerializeStep(const NimbleStepsOutSerializeLocalParticipants* participants, uint8_t* buf,
                              size_t maxCount);
 
-static const int NimbleSerializeMaxRedundancyCount = 10;
+static const size_t NimbleSerializeMaxRedundancyCount = 10;
 
-int nbsStepsOutSerializeCalculateCombinedSize(size_t participantCount, size_t singleParticipantStepOctetCount);
+size_t nbsStepsOutSerializeCalculateCombinedSize(size_t participantCount, size_t singleParticipantStepOctetCount);
 
 #endif
